fix: 修复自助餐列表删除判断

+++ jjj_food_chain/app/shop/model/buffet/Buffet.php
@@ -235,6 +235,10 @@ class Buffet extends BuffetModel
      */
     public function setDelete()
     {
+        if ($this->getCanDelete($this)) {
+            $this->error = '自助餐正在使用中，不可删除';
+            return false;
+        }
         return $this->destroy(['id' => $this['id']]);
     }
